Make search result limit configurable
All search results were previously limited to 10 items and to change this
you would have to modify the PHP files in 3 different places.
This patch adds 'search_limit' as a config item to config/config.php
which will make it easier for customers to modify this value.
Future plans is to make this value configurable from the GUI but this
is a start.

Signed-off-by: Per Åsberg <per.asberg@op5.se>