helpers/nagioscmd: Massage some parameter values when building commands
When building the command string we feed to Nagios, we normally want
timestamps to be in unix epoch style, but that's fairly inconvenient
for users to type. As a result, we need to massage the parameters we
get back from the user and parse it into a proper timestamp.

This patch adds basic support for doing just that by using the fairly
stupid php function strtotime(). It doesn't always get all formats
right, but it's accurate enough that it should suffice for now.

Signed-off-by: Andreas Ericsson <ae@op5.se>