helpers/nagstat: Convert a date format string back to a timestamp
Since we now pass readable date/time strings for commands, we need to
be able to convert them back to timestamps before we submit the actual
command to Nagios. This is as simple as using strtotime() but the date
format for us dates differs from what strtotime can handle, hence we
need this method.

Signed-off-by: Per Åsberg <per.asberg@op5.se>