Remove the noarch tag from building
Since we've got apr_md5_validate (a compiled program) tucked in this
project, it's really not a noarch program. Worst case scenario is that
the package is built on 64-bit systems and doesn't run at all on i386
ones, or that the hash algorithm implementation breaks due to
differences in memory alignment or something else and equally hard
to predict.

Signed-off-by: Andreas Ericsson <ae@op5.se>