controllers/status: Display view even if no data found
Previously we displayed a separate template if no data was found for
a host- or servicegroup. Since we want Ninja to behave like Nagios,
we now modify this to always display the page as when we have data but
with an informative message stating that no data was found.

This fixes bug #2722.

Signed-off-by: Per Åsberg <per.asberg@op5.se>