auth_import_mysql.php: Import nagios users to mysql database
This script will import users and authorization data as described in
cgi.cfg and only works on mysql databases.
Since it is standalone and only requires the path to the ninja directory,
it could very well be used as a cron job to make the users in ninja to
always be in sync with monitor/nagios.

This issue is  closely related to #2743 since the former solution depended
on an existing valid user in the database before importing from nagios.
The downside on this solution compared to the old one is that we lost
the database transparency and will have to create new install scripts
for each database that we want to support.

Signed-off-by: Per Åsberg <per.asberg@op5.se>