helpers/pnp.php: Replace special chars in service names
PNP does some replacing of certain characters in service_description
before storing the RRD and xml files. Ninja should do the same in order
to detect if the correct files exists.
This patch urldecodes the service parameter before substituting '/' and
' ' with '_'. This way we seem to find all files that we previously missed.

This fixes bug #2733

Signed-off-by: Per Åsberg <per.asberg@op5.se>