views/extinfo/index.php: Modified extinfo template
The extinfo template didn't present the host_name information very clear
when viewing info on a service. This patch re-orders some of the info to
make it easier for users to identify the host_name info. Also removed
the ucfirst() that was attached to the host_name string - we shouldn't
modify the data when displaying it.

Signed-off-by: Per Åsberg <per.asberg@op5.se>