helpers/nagioscmd: Convert floatval() duration to seconds
When scheduling flexible downtime for hosts and services, we used
to provide Nagios with a floating point number enumerating the
number of full hours the flexible downtime should last, but Nagios
rejected this number since it wants a count of seconds instead.

This patch massages the 'duration' parameter so that Nagios sees
it as seconds instead.

This should fix issue #2786.

Signed-off-by: Andreas Ericsson <ae@op5.se>