common.js + jquery.qtip.js: Use qtip to render help texts
We use the very flexible JQuery plugin qtip to fetch and display our
help texts. Qtip may be used for lots of other things and is really
simple to use.
This (on common.js) is where we define the look of the help text box
and also what animated load image to use.

Signed-off-by: Per Åsberg <per.asberg@op5.se>