helpers/help.php: Added help text helper
This new helper takes care of rendering a help text icon and the
required code to make inline help work in Ninja.

Help texts are stored in each controller to provide translation
and to keep them logically connected to where they are needed.
The individual help text is fetched when a user clicks on the icon
by an ajax call and is displayed using the JQuery qtip plugin.

Signed-off-by: Per Asberg <perasb@op5.se>