controlers/ajax: Fetch widget settings on Ajax refresh
If the settings for a widget has been updated in the GUI we also want
this change to be included in the next ajax update.
Since the page parameter isn't passed in the ajax update call, this
patch tries to figure out the current page to fetch the correct
settings for the widget.

Signed-off-by: Per Åsberg <per.asberg@op5.se>