install_scripts/auth_import_mysql: Flip $overwrite to enable update
The htpasswd_import class variable $overwrite is by default set to false which
makes it impossible to update the database with new values in case this
script is run after the install. Since we don't have any problem having
this variable set to true on install and that it might make people run
this script manually after the install, we make this change.

Signed-off-by: Per Åsberg <per.asberg@op5.se>