moved the #menu div down 67 pixles to the scrollbar right
the #menu div previously started at the very top and the
first element in the menu had a 67 px top margin. this to be able
to assing the menu-height to 100% without using javascript. now
that we'll be using css-scroll the menu needs to start 67 pixles
down not to hide the top of the scrollbar behind the top panels.
had to adjust some in the javascript as well to adjust to this
change.

Signed-off-by: My Appelgren <my@op5.com>