controllers/command: Check if user is authenticated to submit commands
This patch adds checks against all the parameters available for determining
if a user is authorized to submit commands to nagios.
It involves checks against contact.can_submit_commands and all the
authorized_for_{system_host,service}_commands in cgi.cfg.
It also checks if the user has access to the object (host/service) the
command concerns.

Signed-off-by: Per Åsberg <per.asberg@op5.se>