added clickable icon for displaying version info
also added a name tag to the menu-links to make the menu scroll to
it's right position when a link is clicked. the problem is that
this information needs to be passed along when clicking away within
the pages as well. there is no support for this yet, but i'll look
into it.

Signed-off-by: My Appelgren <my@op5.com>