vendor/html2ps: Introducing html2ps
This is THE biggest library included in Ninja and will hopefully be
removed in a (very) near future. It's advantage is that it is capable
of creating a PDF directly from a HTML page. The disadvantage is that
is slow - very slow.

Don't rely on this one as it will (hopefully) be removed!

Signed-off-by: Per Åsberg <per.asberg@op5.se>