models/reports: Don't check for boolean false for result sets fr db
When kohana returns a database result, we usually get some data about the
query returned if the model doesn't take care of this for us. To make sure,
we should always check count($result). This will work if the model is already
returning false on no data or just returning the query result unfiltered.

In this case we lost all the information about the timeperiod returned from
database because of this faulty check.

Signed-off-by: Per Åsberg <per.asberg@op5.se>