reports/js/common: Make report name editable
This involves creating a drop-down list with all saved reports depending
on in what context we are (avail/sla). This in turn requires a conversion
of the report type id to the identifier equivalent. In order not to hard-code
this and at the same time make the readable name (Availability report) of the
report type translatable, we have to go through creating a json string and
then parse this back to an array in javascript. This information is then
used to determine which javascript array to use when rendering the drop-down.


Signed-off-by: Per Åsberg <per.asberg@op5.se>