models/reports: Add support for 'cluster_mode' reports
A report taken in cluster mode checks for the best state of all the
objects instead of the worst. This is handy when creating reports
for clustered services, such as databases, where the actual service
has been working as good as the best member of the group.

Tests are added to make sure this functions properly.

This is the backend part of the fix for issue #1865.

Signed-off-by: Andreas Ericsson <ae@op5.se>