summary: Only fetch group members when requested
Previously we fetched all available hostgroups when page loaded, dispite
the fact that the user could be interested in a standard report (which
doesn't use this information). This patch moves the step where we fetch
group members, to when the user clicks on the custom report radio button.
Checks are also made to prevent us from fetching the data if the user
keeps toggling between the 2 report types.

Signed-off-by: Per Asberg <perasb@op5.se>