controllers/reports: Skipe some code when generating pdf
We don't need some of the data when generating a PDF but some of the
data needs to be assigned to pdf_data array for later use in _pdf().
This is a very large commit due to quite a lot of indentation changes
to make the code easier to read.

Signed-off-by: Per Asberg <perasb@op5.se>