models/reports: Let callers choose db fields for alert summaries
Since 'output' is a TEXT field, it requires an extra disk hit, which
quickly kills performance. By letting the callers choose which fields
to get, we can avoid that for the "Top Alert Producers", since they
don't need that information anyway.

This provides nearly a 50% performance increase for the top alert
producers case.

Signed-off-by: Andreas Ericsson <ae@op5.se>