models/reports: Avoid bitshifts and logical AND in sql queries
Using such constructs means the query optimizer can't make use of
indices, so we have to do a full table scan more or less every time.

Some indications show that it might be faster to obtain some results
with two queries instead of one, but I don't have time to investigate.
Should we run headlong into performance issues later on, it might be
worth looking in to then.

Signed-off-by: Andreas Ericsson <ae@op5.se>