controllers/default; Added cron() method for scheduled reports
When running the cron script for sending scheduled reports, we need
access to Ninja as a valid user. By calling this method (only from cli)
the config option cli_access is checked and a valid user is retrned
(either from the config file or from the database) and the cron() method
in the reports controller is called with the supplied period string (Daily,
Weekly, Monthly).

Signed-off-by: Per Asberg <perasb@op5.se>