models/reports: Update Reports_Model to use PDO for all its real work
The _self_check() function still doesn't, but that's of no moment since
it only executes a very simple query. The timings prove beyond a doubt
that we took the right path here.

With the patch:
All 85 tests passed. Hooray!
0.42user 0.18system 0:00.66elapsed 90%CPU (0avgtext+0avgdata 0maxresident)k
0inputs+0outputs (0major+19269minor)pagefaults 0swaps

Without the patch:
All 85 tests passed. Hooray!
0.61user 0.25system 1:17.81elapsed 1%CPU (0avgtext+0avgdata 0maxresident)k
0inputs+0outputs (0major+19727minor)pagefaults 0swaps

The patched version is 117.89 times faster than the unpatched one.

It should be noted that the tests being run contains one comparatively
huge testcase that is not committed to version control.

Signed-off-by: Andreas Ericsson <ae@op5.se>