models/reports: Accept 'start_time' and 'end_time' in human-readable formats
We only accept values that can be converted using strtotime() though.
In practice, that's most of the dates. iso8601 seems to be the
preferred interpretation with ambiguous dates in the form xxxx-xx-xx
which could otherwise mean yyyy-mm-dd or yyyy-dd-mm.

I have no idea how it parses xx-xx-xx dates, which will be ambiguous
in some way until 2031, but that's hardly interesting.

This also allows us to get rid of the datetime_to_timestamp() function,
which had the same ambiguity problems (ie, we didn't really know).

Signed-off-by: Andreas Ericsson <ae@op5.se>