controllers/trends: Make event duration bars a config option
The new trends reports is quite different to the old one in many ways.
One major thing is that we now have the possibility to display state history
for several objects (of the same type) at once. The drawback is that it can
get pretty crowded and hard to read. This is why the default option is to
hide the event duration bars. The use nevertheless has the possibility to
override this setting in the setup.

Signed-off-by: Per Asberg <perasb@op5.se>