models/reports: Use a single database instance
When creating a large report with sub-reports, we used to have one
database connection for each sub-report and one connection for the
master report. Since we really only need one for all the reports we're
far better off re-using the one from the master-report in all the
sub-reports.

This patch does just that, solving the case where the database layer
can run out of connection handles if we try to create really large
reports.

Signed-off-by: Andreas Ericsson <ae@op5.se>