models/reports: Support filtering by objects
This patch adds object-based filtering to the alert summary query
builder. This is necessary in order to avoid fetching more data than
we actually need, and lets the database do what it does best (ie, find
only the interesting items) so we don't have to implement it in the
application logic.

Signed-off-by: Andreas Ericsson <ae@op5.se>