controllers/summary: Add method _print_alert_totals_table()
This method prints one alert totals table, given the result set and a
translation array for the various state id's, as well as a topic and a
totals array. Since all alert totals tables look more or less the same,
we can get away pretty cheaply here by reusing a single function for
quite a lot of the various alert total report types.

Signed-off-by: Andreas Ericsson <ae@op5.se>