models/reports: Make queries easier to debug
Since the database layer reports errors based on line-number, we can
quite easily make query errors easy to spot by the simple expedient of
line-breaking our queries and making sure they're printed with
linenumbers in case we run into any errors. Mentally adding one to
the numbering scheme is a manageable task, so we use a simple
explode("\n", $query) to print it.

Signed-off-by: Andreas Ericsson <ae@op5.se>