helpers/reports: Added method js_strings()
Quite a lot of the translated javascript strings are common for all the
reports (reports,histogram,trends,summary) and putting them in this
helper minimizes the risk of forgetting any of them. Failing to add them
will result in javascript errors and broken functionality.

This commit contains the helper method and all the patches made where it
is used.

Signed-off-by: Per Asberg <perasb@op5.se>