{histogram,reports,summary,trends}: consolidated javascript files
Previously we had one javascript for each of the reports (histogram,reports,
summary,trends) which made debugging a pretty time consuming procedure.
The solution to this is to put the common javascript code in
reports/common.js and keep the parts that are unique to the report type
in their respective js directory.

Signed-off-by: Per Asberg <perasb@op5.se>