reports: Make direct link to schedule reports tab work
We would like to have a direct link from the menu to where we
display/edit/add schedules. The simple solution for this is to watch
for a custom GET variable called 'show_schedules'.
Whenever we find this variable we set a special javascript variable
that changes the default selected tab.

Signed-off-by: Per Asberg <perasb@op5.se>