controllers/{extinfo,status}: Correct page links
Some of the page links from status and extinfo controller were created
before we actually knew where to point them.
Since all cgi's has been converted we can now substitute the page links
with working ones.
Still missing for notification links though since that controller isn't
yet able to to accept links for specific host or services etc.

This is a fix for issue 3066 and a prtial one for issue #3069

Signed-off-by: Per Asberg <perasb@op5.se>