controllers/command: Refactor _is_authorized_for_command()
Previously it regarded an empty $param struct as something erroneous,
but that's not entirely true since all system commands always have
empty such structs.

Secondly it was overly eager to fetch configured hosts and services
from the database when it's quite easy to check if we can bypass
doing that entirely, given the right cgi.cfg parameters for the
user in question.

It was also unnecessarily deeply indented. We remove a lot of that
by issuing early return() statements, avoiding a lot of else{}
blocks.

Incidentally, this fixes issue #3059.

Signed-off-by: Andreas Ericsson <ae@op5.se>