models/nagios_auth: Map the cgi.cfg parameters to their proper variables
Previously, authorized_for_all_host_commands was mapped to
authorized_for_all_hosts, which was both bogus and confusing. The same
was in effect for services.

This patch updates it to also use the command_hosts_root and
command_services_root variables, as well as fixing the erroneous
mapping, and also takes care of some DWIMery to match the docs at
http://nagios.sourceforge.net/docs/3_0/configcgi.html

Signed-off-by: Andreas Ericsson <ae@op5.se>