pagination.js: Make submit save GET params
The previous version of this code demanded that the user clicked on the
'Go' button to trigger it. Since most users will change the value and the
press enter we won't get the extra parameters anyway.

The solution is to bind the code to be triggered on the submit event.

Signed-off-by: Per Asberg <perasb@op5.se>