Only count unique entries when creating reports
The idea is that the database should only hold unique entries, but
it seems this doesn't always work. Making sure this is the case
when reading the data filters out quite a lot of duplicate entries
it seems, since Top Alert Producers numbers go down by a LOT with
this patch enabled.

This, in a way, fixes issue #3183.

Signed-off-by: Andreas Ericsson <ae@op5.se>