views/reports/js/reports: Make it possible to save a report when generated
This worked pretty much before but since we had to rework things a bit for
SLA reports there were quite some more things to be checked etc.
One things that makes this tiresome is the fact that fancybox is kind of
stupid making a copy of the part it displays. This makes it tiresome to
work with forms as you end up with a new instance of every form you put
in a fancybox wich in turn forces you to write specific code for fancybox
forms.

Signed-off-by: Per Asberg <perasb@op5.se>