controllers/reports: Prevent javascript from messing up reports
When date + time is passed from javascript it is concatenated into one
form field. Problem is that we have some case when we get 'undefined'
from javascript which messes up the report. This will be solved in
the javascript as well but his is just to make sure we don't fail in
any way.

Signed-off-by: Per Asberg <perasb@op5.se>