reports: Re-worked the schedule add/delete functions
Since we display schedules in 3 diferent places, we needed to update
the javascript part and templates to make it possible to add/delete them
in the various places.

This includes hiding/showing messages that there are no available schedules
for a certain type and to show/hide table header fields depending on if we
have any schedules or not.

Signed-off-by: Per Asberg <perasb@op5.se>