controllers/reports: Pass saved report info to template
We used to fetch info and manipulate it and update the form through
javascript but since there's no need for this as the page is being
reloaded anyway we might as well pass the data by assigning it to template
variables. This has the benefit of using PHP power to convert dates and
such instead of trying to solve it by javascript.

Signed-off-by: Per Asberg <perasb@op5.se>