models/user + controllers/default: Move auth sync from user model
We used to have the authentication data to be synced against cgi.cfg
and htpasswd.users when user was authenticated. This had the obvious
drawback that a valid user had to login before an updated or added
user coould log in.

This patch moves the syncing to after we have validated that the user
entered some login data that seems OK but before we actually validated
the user.

This fixes issue #3168

Signed-off-by: Per Asberg <perasb@op5.se>