reports: Re-make SLA date selection
As we only support SLA reports by month, the SLA reports shouldn't break
between months. This patch fetches data from the database limiting the
start- and end_dates to dates actually present in the database.

We still need to fix the population of a saved report to make issue 3128
solved.
Partial fix for issue #3128

Signed-off-by: Per Asberg <perasb@op5.se>