Make loading of custom SLA report work with new date selection
Since date selection in SLA reports works a little differently than in
avail reports, we also have to make it possible to load the saved report
properly. One problem here is that the year/month information in the
setup form is fetched on load, forcing us to check that the values are
present before trying to set anything as selected. By using some JQuery
this is no big chore and the selection is delayed one second each time
we notice that the form isn't ready.

This is the last fix for issue #3128

Signed-off-by: Per Asberg <perasb@op5.se>