Add Makefile to simplify running tests and building documentation
Running the testcases is a bit of a pain since it lacks a short-command.
Adding the Makefile means we have something that will always fail in
case a test-case fails, and it makes it easy to add specific tests for
specific areas. Currently, 'make test' runs all available tests, which
so far consists only of report tests, which can be explicitly selected
by running 'make test-reports'.

'make Documentation' and 'make docs' will build the doxygen generated
documentation, with proper dependencies to ensure we never rebuild it
needlessly.

Signed-off-by: Andreas Ericsson <ae@op5.se>