models/reports: Support keeping sub-report logs intact
With this commit we add the option of maintaining subreport logs,
which act as single object logs. This is a good thing, since those
logs are always sequential by object and always belong to a single
object, as opposed to the full group-log which can be hard to
interpret at times.

Signed-off-by: Andreas Ericsson <ae@op5.se>