models/system: Added methods to fetch and filter data from status.log
Since some performance information is missing in merlin and some might
find it valuable, we can now fetch this information from status.log.
This is ofcourse a more time consuming process than fetching it from
the database but since it is only done for the Performance Info page,
we can very well support this way of fetching the data.

This is a partial fix for issue #3255

Signed-off-by: Per Asberg <perasb@op5.se>