status/{host/service}: Don't query every object for comment count
We use the new Comment_Model::count_comments_by_object() method to
fetch an array with objects that has a comment and use this to display
the comment icon or not.
Doing things this way will require only 1 query per object type
(host/service) instead of one for each object.

Signed-off-by: Per Asberg <perasb@op5.se>