models/reports: Revamp updating of the state log
This patch makes the log-entries for a group of objects, ad-hoc or
configured, reflect the state of that entire group properly. Logs
for each individual objects are not stored in the group's log.

The generated output can certainly be improved, but this patch will
at least enable us to make proper trends reports as it is.

Signed-off-by: Andreas Ericsson <ae@op5.se>