trends: Created new trends report
Our new trends reports wasn't as good as we hoped so we are now introducing
a new report. It is pretty similar to the old trends report in the cgi
but we are confident that it will suffice. It has the benefit that it is
possible to use together with the avail report as well.

Signed-off-by: Per Asberg <perasb@op5.se>