reports: Fetch field value for a saved report
If the user tries to update a filed with a value that is not accepted,
we need to be able to let them restore it from the saved data.
This patch creates an ajax call that fetches the data and assigns it to
the faulty field in the page.

Fixes issue #3277

Signed-off-by: Per Asberg <perasb@op5.se>