controllers/reports: Move css to template collection part
Since we usually asign all the css resources early in the controllers,
the part that assigns the added css files to the master template is also
placed at the top of the method. Here we want to add some resources rather
late in the method depending on if we are displaying a report for a single
object or not and hence we need to add the css resources later. This forces
us to place the css collection at the bottom of the controller method.

Signed-off-by: Per Asberg <perasb@op5.se>