helpers/showlog: Pass --cgi-cfg to the showlog helper app
We needn't bother with passing all the logfiles (which may not even
be there) to the showlog app anymore, since it now supports parsing
everything it needs from cgi.cfg and nagios.cfg. Since the location
of nagios.cfg can be inferred from cgi.cfg it's enough to pass the
location of the cgi config file and just leave it at that.

This finalizes the fix for issue #3235.

Signed-off-by: Andreas Ericsson <ae@op5.se>