views/reports/js/common.js: Make custom report period work for all reports
We used to have a check that we were using custom and avail report for some
checks to be executed. This was fundamentally wrong since the only time
we shouldn't perform the checks is when we're editing an SLA report.
Changing the check value makes custom report periods work on all our
reports.

Signed-off-by: Per Asberg <perasb@op5.se>