helpers/showlog: Re-arrange command-line slightly
We put all the html options last to make it easier to cut them out
when one has to debug the showlog application. We also put --reverse
second to last. It was the most recent culprit and it's usually
useful to make sure two runs, with and without --reverse, at least
produces the exact same amount of lines, even if they're not the
same ones.

Signed-off-by: Andreas Ericsson <ae@op5.se>