helpers/showlog: Limit output to 2500 lines by default
If we let the showlog helper application print all the logs we
run a very real risk of hitting the PHP memory barrier, as well
as crashing our users' browsers when they try to load the huge
page.

If the user specifies a start or an end-date, or manually sets
a limit, we add no additional limits.

Signed-off-by: Andreas Ericsson <ae@op5.se>