controllers/command: Fix host downtime scheduling
Previously we used to send two (or three, if 'services too' was
checked) commands into Nagios, setting first the host itself to
be in scheduled downtime and then once more in case something
was selected for child hosts. This didn't fly too well, since the
command to schedule downtime for child hosts too also schedules
downtime for the parent host, which in essence ends up in two
levels of downtime.

This patch corrects the problem, blocking the original command
if we're also scheduling downtime for child hosts.

This fixes issue #3285.

Signed-off-by: Andreas Ericsson <ae@op5.se>