reports/schedules: Make it possible to send an already scheduled report
Previously we could only send scheduled reports that were scheduled in
one sequence: create, save + schedule. Any report that was saved and then
reloaded, was never ever possible to send by email.
This patch changes this, adding the "send now" button to all generated
reports when viewing the schedules for it.

Signed-off-by: Per Asberg <perasb@op5.se>