Detect stale data in Ninja controller
We want to alert user in case the data is too old in the database so they
can take actions, we do this in Ninja controller and use the config
option config.stale_data_limit. The ninja class member $stale_data is set
to the diff between current time and last_alive (in seconds).

Signed-off-by: Per Asberg <perasb@op5.se>