Check for data freshness when page refresh is disabled
Since some of the pages (like tac) in Ninja has refresh disable we need
some other way of validating that data is not stale.
This is done through the refresh helper which adds an ajax call to be
run in the same interval as defined in config.stale_data_limit.

Signed-off-by: Per Asberg <perasb@op5.se>