core/Kohana: Don't register error- and exception handler in cli mode
Doing so causes the screen to be filled with nonsensical html code
which isn't exactly user-friendly. This patch makes the framework
use the default php error-handler (or the one from any installed
modules) in case we're running in cli mode, which is a lot better
for debugging tests designed to run in command line mode.

Signed-off-by: Andreas Ericsson <ae@op5.se>