models/saved_reports.php: Don't save {start,end}_time when report period != custom
Saving {start,end}_time for reports that already has a report_period that is not custom,
will make some checks in the reports controller to fail. This is because the time values
will be passed as '2010-05-10' and since it is stored in an integer field it will be
saved as 2010 and be interpreted as something like 1970-01-01, 00:33 and be printed
in the custom date field when/if a user switches to custom report period.

Signed-off-by: Per Asberg <perasb@op5.se>