controllers/command.php: MAke command controller handle multiple objects
Since we are implementing multiple object selection for issuing commands, we
need to make the command controller to handle this.

This is done partially by wrapping the call to nagioscmd::build_command()
and to check if the passed host or service parameters are arrays.

This is a partial fix for issue #1525

Signed-off-by: Per Asberg <perasb@op5.se>