Search: Added more search features
These patches adds AND and OR search to Ninja as well
as a possibility to change the limit.

Now it's possible to search for services on a subset
of hosts. By typing:
h::linux OR windows AND s:ping OR http limit=100
a search will be performed for services called something 
like ping or http on hosts called something like linux
or windows and the result will return a maximum of 100
rows.

This fixes issue #3391

Signed-off-by: Per Asberg <perasb@op5.se>