Add filter for report objects
When selecting objects for a report, it is sometimes
pretty time consuming to scroll through a list of hundreds
(or even thousands) of objects to find what you are looking 
for.

This patch adds a filter box to the avail and SLA report
setup page that amends this problem. Some styling still 
left to be done and this should also be added to some of
the other report types (Trends, Custom Alert Summary).

This fixes issue #2074

Signed-off-by: Per Asberg <perasb@op5.se>