controllers/reports: Stash report parameters for later
We would like to keep the used parameters for each
report in order to display backlinks when drilling
down and a direct link to each report.

This patch adds the method _stash_params() that
takes care of this for us and stores the information
in the current session.

Partial fix for issue #3644 and #3631

Signed-off-by: Per Asberg <perasb@op5.se>