controllers/reports: Increase trends graph
The trends graph showed in avail reports were
somewhat smaller than the one in trends and this
somehow caused some of the background images to
be too small - so small that you couldn't see them.

This small patch makes the trends graph in the 
avail report to have the same height as in trends.

Signed-off-by: Per Asberg <perasb@op5.se>