Updated group pages
The {host,service}group pages needed to be rewritten
due to bad response times. SQL queries are modified
for better performance and the links to these pages 
are changed to include a pagination variable. This
since the default pagination is 100 items and one 
item in these cases are a host- or a servicegroup
which usually leads to large result sets. Therefore
the initial page when reached from a link only
contains 10 groups.

Signed-off-by: Per Asberg <perasb@op5.se>