Make Auth work with integer usernames and with @
The Ninja Auth library didn't work properly for
users with usernames beginning with integers or
username that had an '@' sign (like an email
address). This was due to the over-intelligent
(or maybe stupid) Auth ORM models trying to detect
the type of variable passed to ORM::factory().

This patch tells Ninja Auth to always look in the
username field regardless of what the username
looks like.

This fixes issues #3714 (again) and #3276

Signed-off-by: Per Asberg <perasb@op5.se>