reports: Create temp directory for temporary reports files
When creating PDF reports using tcpdf, we need to
create some files before adding them to the PDF
and also the PDF file before sending. This patch 
makes all those files end up in temp directory
that is removed once the report is sent. By 
letting the scripts create this directory, we also
get the possibility to delete it afterwards which
previously was a problem.

This fixes issue #3723

Signed-off-by: Per Asberg <perasb@op5.se>