Display all comments for an object
The comment section was pretty messed up since we
tried to do some magic to display user comments
for scheduled downtime entries. This had some 
pretty weird side effects resulting in, well
confusion to be honest.
This patch makes us display all comments and 
downtime comments for each object. 
As a (hopefully) good side effect we now give the 
user the possibility to delete a scheduled
downtime from the extinfo page as well instead of
having to switch view to the schedule downtime
page.

This fixes (again) issue #3529

Signed-off-by: Per Asberg <perasb@op5.se>