controllers/status: Ignore 'true' as value
When sorting and changing pagination from host- or
status details page, the jquery.query plugin
treats empty GET variables as having the value
true. This is usually OK but in the status
controller we do some gaic using this field in
order to display correct oage title.
Because of this the hoststatustypes and
servicestatustypes variables are now handled
specifically for this issue.

This fixes issue #3776

Signed-off-by: Per Asberg <perasb@op5.se>