Show hostgroups when only partially authorized
The Nagios default is to only show host- or 
servicegroups for a limited user when they have 
access to all included hosts or services. 
This patch makes this configurable and the only 
change required is switching the appropriate flag 
in config/groups.php

It is recommended for users to add a copy of this
file in config/custom/ to make sure changes aren't
overwritten on upgrade. 

This fixes issue #2221

Signed-off-by: Per Asberg <perasb@op5.se>