Limit access to extinfo group commands
Since it is now possible to view information about
groups which a user is only partially authorized,
we here introduce an extra check that is made 
before displaying any group command links on the
extinfo page. This is done by comparing nr of
authorized host/services with member of the 
current group and to display an error page if 
those doesn't match.

Signed-off-by: Per Asberg <perasb@op5.se>