ajax controller: Cleanse parameters before passing them to pnp
Otherwise popup graphs for services and hosts with slashes, spaces,
colons and backslashes won't work. This uses the same regex to
clean as pnp itself does, so it should be safe. We obviously implement
it as a method in the pnp helper class, so that it's reusable in other
places.

Signed-off-by: Andreas Ericsson <ae@op5.se>