Optimized group overview pages
This patch speeds up the host- and servicegroup
overview pages quite dramatically. The key here
is that we now let the database do more of the 
heavy lifting and that we let the view (template)
call the model directly to avoid large array
structures in PHP and instead loop over the
db result directly.

Signed-off-by: Per Asberg <perasb@op5.se>