js_header.js: Add support for external resources
Since Kohana tries to be smart (?) and are adding
.js to every file that doesn't end in this, we
here add support for external files which simply
bypasses this check and echoes out the correct
script tag.

This doesn't solve a possible problem if anyone
would want to use parameters to local javascript
files but we will handle this when/if it occurs.

Signed-off-by: Per Asberg <perasb@op5.se>