Make popups configurable
This patch makes the use of popups for comments and
PNP graphs a little more configurable by making
it a user option under My Account.

It is now possible to turn it off and to control
the delay before the pop-ups are shown.
Default values are ON and 1500ms for the delay.

This fixes issue #3837

Signed-off-by: Per Asberg <perasb@op5.se>