Make sure we have enabled scheduled summary reports
It has been reported that some systems doesn't
have summary as a valid scheduled report even
though the table for saving info about those
reports exists. Given that the record for enabling
the summary reports are created in the same sql
file makes things pretty weird. 

This patch tries to work around this by checking
for the 'summary' value and adding if it doesn't
exist. 

Signed-off-by: Per Asberg <perasb@op5.se>