models/host: Make sure we don't fail on serviceprops
If serviceprops are passed to status/host we will
end up in a situation where we are trying to fetch
info on services and since this table isn't 
included in the FROM part we will end up in an
ugly Kohana error. 

This patch takes care of this even though we 
shouldn't even end up here (status/service should
have been used instead).

Signed-off-by: Per Asberg <perasb@op5.se>