Added context menu for status pages
This patch includes the necessary changes to add
context menus for status/host and status/service
as well as status/{host,service}group_overview.

The menu is somewhat context sensitive and sets
the correct menu values depending on object status
and properties. This is unfortunately not true for
the group_overview page since this would require
fetching a lot of extra data, hence it always
displays all the available menu alternatives
regardless of object (host) state.

Signed-off-by: Per Asberg <perasb@op5.se>